todos/serializers.py
from rest_framework import serializers
from .models import ToDo


class ToDoSerializer(serializers.ModelSerializer):
    id = serializers.UUIDField(read_only=True)

    class Meta:
        model = ToDo
        fields = ["id", "title", "task", "finished_time", ]
